% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{gen_url_pm}
\alias{gen_url_pm}
\title{Generates URL for pubmed API}
\usage{
gen_url_pm(
  searchterm,
  startdate = as.character(Sys.Date() - 365, "\%Y/\%m/\%d"),
  enddate = as.character(Sys.Date(), "\%Y/\%m/\%d")
)
}
\arguments{
\item{searchterm}{text of the query}

\item{startdate}{from date appeared online (default = 1 year ago today), format YYYY/MM/DD}

\item{enddate}{to date appeared online (default = today), format YYYY/MM/DD}
}
\description{
Generates URL for pubmed API
}
