% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{get_pm}
\alias{get_pm}
\title{Pubmed all steps}
\usage{
get_pm(
  searchterm,
  startdate = as.character(Sys.Date() - 365, "\%Y/\%m/\%d"),
  enddate = as.character(Sys.Date(), "\%Y/\%m/\%d")
)
}
\arguments{
\item{searchterm}{text of the query}

\item{startdate}{from date appeared online (default = 1 year ago today), format YYYY/MM/DD}

\item{enddate}{to date appeared online (default = today), format YYYY/MM/DD}
}
\value{
a tibble of results (8 cols if articles present, 0 if not)
}
\description{
Pubmed all steps
}
