% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{get_results_pm}
\alias{get_results_pm}
\title{Gets results from pubmed API}
\usage{
get_results_pm(searchurl)
}
\arguments{
\item{searchurl}{URL with search query}
}
\description{
Gets results from pubmed API
}
