% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{get_results_springer}
\alias{get_results_springer}
\title{Fetch results from Springer API}
\usage{
get_results_springer(searchurl, searchterm)
}
\arguments{
\item{searchurl}{URL with the search query}

\item{searchterm}{original search term (for final filtering)}
}
\description{
Fetch results from Springer API
}
