% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{get_springer}
\alias{get_springer}
\title{Springer all steps}
\usage{
get_springer(
  searchterm,
  datefrom = as.character(Sys.Date() - 365),
  dateto = as.character(Sys.Date()),
  apikey = Sys.getenv("SPRINGER_API")
)
}
\arguments{
\item{searchterm}{text of query}

\item{datefrom}{earliest date added}

\item{dateto}{latest date added}

\item{apikey}{Springer API key}
}
\description{
Springer all steps
}
