% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{search2filter}
\alias{search2filter}
\title{Turn a search term into a dplyr filter condition}
\usage{
search2filter(searchterm)
}
\arguments{
\item{searchterm}{Search term (including AND, OR, NOT, quotes and brackets where required)}
}
\value{
A string to pass to a dplyr filter
}
\description{
Turn a search term into a dplyr filter condition
}
